Fix link failure when building olvwm due to missing 'cfree'
When building olvwm on Ubuntu 20.04, linking fails due to a missing 'cfree'
As per the man page on Ubuntu 20.04 for cfree(3):
==begin quote==
This function should never be used. Use free(3) instead. Starting with
version 2.26, it has been removed from glibc.
==end quote==

So I followed the manpage and replaced all calls to cfree(3) with calls
to free(3)